requirements.txt
boto==2.38.0
cycler==0.9.0
Cython==0.23.4
h5py==2.5.0
Keras==0.3.0
matplotlib==1.5.0
numpy==1.10.2
Pillow==3.0.0
pydub==0.16.0
pyparsing==2.0.6
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
requests==2.9.1
scikit-learn==0.17
scipy==0.16.1
six==1.10.0
Theano==0.7.0
wheel==0.24.0
